completed jwt-extended lessons

+++ resources/item.py
@@ -1,6 +1,5 @@
-import sqlite3
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims, jwt_optional, get_jwt_identity, fresh_jwt_required
 from models.item import ItemModel
 
 
@@ -9,7 +8,7 @@ class Item(Resource):
     parser.add_argument('price', type=float, required=True, help="This field cannot be left blank")
     parser.add_argument('storeId', type=int, required=True, help="Store Id field cannot be left blank")
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
@@ -17,6 +16,7 @@ class Item(Resource):
         else:
             return {'message': 'item not found'}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': 'Item with name: {} already exist'.format(name)}, 400
@@ -31,7 +31,11 @@ class Item(Resource):
             return {'message', 'An error occurred inserting an item'}, 500
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {"message": "Admin privilege required."}
         item = ItemModel.find_by_name(name)
         if item:
             item.delete()
@@ -51,6 +55,8 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {"items": [item.json() for item in ItemModel.query.all()]}
+        user = get_jwt_identity()
+        return {"items": [item.json() for item in ItemModel.find_all()]}
         # return {"items": list(map(lambda item: item.json(), ItemModel.query.all()))}
